商品属性SKU管理接口

+++ model/src/main/java/com/atguigu/gmall/product/entity/SkuSaleAttrValue.java
@@ -29,7 +29,7 @@ public class SkuSaleAttrValue implements Serializable {
     /**
      * spu_id(冗余)
      */
-    private Integer spuId;
+    private Long spuId;
 
     /**
      * 销售属性值id
